pyproject.toml
[project]
name = "fyp-cluster"
version = "0.1.0"
description = ""
authors = [
    {name = "SyntaxaR",email = "83219306+SyntaxaR@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "st7735>=1.0.0",
    "pillow>=12.1.0",
    "gpiod>=2.4.0",
    "gpiodevice>=0.0.5",
    "onnxruntime>=1.23.2",
    "numpy>=2.2.6",
    "opencv-python>=4.13.0.90",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.90",
    "scipy>=1.15.3",
]
